+++ RELEASE-NOTES.md
@@ -55,8 +55,12 @@ Initial version with:
 ## [0.10.1](https://github.com/kisscool62/io.nagurea.smsupsdk/compare/v0.10.0...v0.10.1) (2022-07-17)
 * Fixed document Main exemple + set public accessor on *Service constructors for easier build without spring or dedicated factory
 
-## [0.10.2](https://github.com/kisscool62/io.nagurea.smsupsdk/compare/v0.10.1...v0.10.2) (2022-07-17)
+## [0.10.2](https://github.com/kisscool62/io.nagurea.smsupsdk/compare/v0.10.1...v0.10.2) (2022-08-08)
 * Invoices > Download invoice
+
+## [0.11.0](https://github.com/kisscool62/io.nagurea.smsupsdk/compare/v0.11.0...v0.11.0) (2022-08-09)
+* API Token > Create a token (directly or from subaccount)
+
 
 ### Upgrade Steps
 * [NO ACTION REQUIRED]
@@ -65,7 +69,7 @@ Initial version with:
 * [NO BREAKING CHANGE]
 
 ### New Features
-* Invoices > Download invoice
+* API Token > Create a token (directly or from subaccount)
 
 ### Bug Fixes